random/src/test/java/ga/rugal/amazon/wordsearch/SolutionTest.java
package ga.rugal.amazon.wordsearch;

import java.util.stream.Stream;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

public class SolutionTest {

  private final Solution solution = new Solution();

  static Stream<Arguments> source() {
    return Stream.of(
      Arguments.of(new char[][]{new char[]{'A', 'B', 'C', 'E'}, new char[]{'S', 'F', 'C', 'S'}, new char[]{'A', 'D', 'E', 'E'}}, "ABCCED", true),
      Arguments.of(new char[][]{new char[]{'A', 'B', 'C', 'E'}, new char[]{'S', 'F', 'C', 'S'}, new char[]{'A', 'D', 'E', 'E'}}, "SEE", true),
      Arguments.of(new char[][]{new char[]{'A', 'B', 'C', 'E'}, new char[]{'S', 'F', 'C', 'S'}, new char[]{'A', 'D', 'E', 'E'}}, "ABCB", false)
    );
  }

  @ParameterizedTest
  @MethodSource("source")
  public void testExist(final char[][] board, final String word, final boolean expected) {
    Assertions.assertEquals(expected, this.solution.exist(board, word));
  }
}
